use waves

+++ src/instr.h
@@ -1,4 +1,4 @@
-/* #include global.h */
+/* #include "global.h" */
 
 double instr_env_get(Note* note);
-double instr_get(Note* note, long srate);
+float instr_get(Note* note, long srate);
